[usdImaging, hdx, usdImagingGL, usdviewq] Improve camera and framing API
usdImaging:
- Fix bug when handling multiple camera attributes tied to a single dirty bit.

usdImagingGL:
Previously, UsdImagingGLEngine::SetCameraState was used to set framing state for the active camera, regardless of whether it was a scene camera or the free camera.
The camera state was directly used by the legacy engine; when using Hydra, a lone camera Sprim was updated via HdxTaskController with the active camera state (regardless of whether it was scene camera or free camera).
With the recent addition of a usdImagingCameraAdapter, USD scene cameras manifest as camera Sprims in Hydra, and thus provide a 1:1 mapping b/w the USD stage and Hydra. This allows Hydra tasks that take a cameraPath as a parameter to use a USD camera path, and in the future, get physical camera parameters to simulate camera image effects such as depth of field.

- Add explicit methods to set a USD camera (SetCameraPath), set a viewer window policy (SetWindowPolicy) and set the viewport size (SetRenderViewport).
- Reworked existing SetCameraState method to take view and projection matrices. The viewport isn't camera state, and is set via SetRenderViewport.
- Add special handling for the legacy engine when a scene camera is active (to extract the relevant camera state from the stage).
- Updated client sites.

hdx:
- Add SetCameraPath(id) to use a USD camera path when rendering.
- Rename SetCameraViewport to SetRenderViewport.
- Rename existing camera state setters to clarify that they update the free camera (SetFreeCameraMatrices, SetFreeCameraClipPlanes).
- Remove SetCameraWindowPolicy since it was unused. The free cam frustum is pre-adjusted and it defaults to the window policy Fit; we can reinstate this method when frustum conformance for the free cam is moved into UsdImagingGLEngine.

usdviewq:
- Add utility method to get active scene camera prim (getActiveSceneCamera)
- Use appropriate method in paintGL when the active camera is a free camera or scene camera.

(Internal change: 1979016)