usd: Introduce UsdStageLoadRules, to support more robust and declarative intent about how payloads should be included in UsdStages.  Also include both LoadRules and PopulationMasks in usd instancing keys, so that different instances can have different load states.  Instances that have equivalent load rules and population masks still share, but those that do not will generate different masters.
With this change, Load/Unload operations on UsdStages now must operate
on the "expanded" stage namespace.  That is, the namespace as it would
be if instancing was disabled.  Said yet another way, the namespace
that's represented by traversing with Instance Proxies enabled.

(Internal change: 1981783)
(Internal change: 1981892)
(Internal change: 1982024)
(Internal change: 1982059)