Remove the "geo" bridge location from prototype paths in Katana. The "geo" location was created as a stub for transitioning from BuildIntermediate to PxrUsdIn as part of building out prototypes under a point instancer. Prior to change 1793604, this bridge location did not have access to PxrUsdIn private data, meaning that it couldn't represent a location coming from USD. To work around this limitation, we created the stateless "geo" location and inserted it into the original hierarchy. However, thanks to change 1793604, the bridge location can now be a USD location instead of a stateless stub. Tell BuildIntermediate to start building just above the prototype (or the "common ancestor" between it and its Looks). The common ancestor will be read as if it were any other USD location.
Since BuildIntermediate now starts building one scope above where it used to, we can't author attrs on the common ancestor in the same way as before. Simply nix the "info.usd.sourceUsdPath" tracking attr, since it is unused. As far as the "instance source" type attr goes, it turns out that using the common ancestor as the instance source instead of the prototype doesn't work in RfK 22, so this change also includes a modification to how we tag instance sources for point instancer prototypes.

When building out point instancer prototypes, identify each prototype as an instance source rather than identifying the common ancestor as an instance source. RfK 21 was more permissive about instance source rules, and allowed an instance array to have geometry.instanceSource entries that did NOT correspond to instance source locations. However, RfK 22 is not as lax, and now expects the geometry.instanceSource entries to be, as the name implies, instance sources.

The classic way to identify an instance source is by changing its type to "instance source". However, this causes problems if the prototype's USD type (and therefore its Katana type) holds special meaning. So, in the cases where the prototype's native Katana type is deemed important, insert an empty group into the hierarchy to hold the "instance source" type. This requires PxrUsdIn to be aware of a new tag stored in the static scene attributes ("insertEmptyGroup"). If this attribute is present, PxrUsdIn will pass the current prim to createChild, as if it were its own child, and early exit, thereby turning the prim's original Katana location into an empty group.

See #834
See #835

(Internal change: 1985251)