[hd,hdSt] texture resource handles for materials
Updated texture resouce binding in HdStSurfaceShader to
add a level of indirection between the shader's texture
descriptors and the underlying GL texture resource IDs.

This is exposed as a concept only within HdSt and does not change
the HdSceneDelegate interface.

This is implemented by registering an HdStTextureResourceHandle
in the HdStResourceRegistry which can be used to resolve
the current texture resource for a given texture handle ID
The texture handle ID is the ID of a texture prim or draw target
attachment.

Using texture resource handles allows some common cases
of texture updates (e.g. animated texture cards) to proceed
without needing to update material parameters.

We still need to propagate texture resource changes to
materials in some cases, i.e. when the texture type has changed
(e.g. UV vs Ptex, etc) or when using bindless textures.

(Internal change: 1993404)