Update UsdPreviewSurface shader to use 'renderType' metadata tagging.  This change follows a previous checkin 1993725 that adds support for SdrPropertyTypes->Terminal, which requires the 'renderType' metadata tag in order for a property to be identified as this new type.
We need to update UsdPreviewSurface so that the terminal typed properties get translated to the correct SdfValueTypeName ('token') when they go through Sdr.
Without this change, the properties that should be terminals will get translated to 'string' types, which are unconnectable to UsdShadeMaterial, whose spec defines terminal properties to be of 'token' type.

(Internal change: 1993866)