Allow "worldtoscreen" image metadata needed for cards drawMode to be represented using different types.
This metadata was previously required to be a vector
of 16 floats. However, this may not be supported by
certain file formats, or may be inconvenient. For
example, OpenImageIO will automatically convert a
float[16] array to a m44f matrix when writing
metadata to a .exr file. We now allow "worldtoscreen"
to be a vector of 16 floats or doubles, or a 4x4
matrix of floats or doubles.

(Internal change: 1995058)