IsPathIncluded results in a coding error if a relative path is provided to it.
IsPathIncluded is a dumb check of paths having no knowledge of the usd stage infrastructure. And during traversal of path to check against exclusion rules, it needs to get parent path. Since for a relative path, parentPath will never end up being an "EmptyPath", the path traversal loop never exists for one of the overrides of IsPathIncluded.
Also added a test to test the appropriate behavior.

(Internal change: 1997263)