Removing 'surface' and 'displacement' from SdrNodeRole
The remaining entries 'primvar', 'texture' and 'field' make sense for individual nodes. Surface, displacement and volume nodes are identified by the connection/relationships from the enclosing material. Having those as roles does not make sense and is not used.

There is only a single instance were this metadatum was authored, which is UsdPreviewSurface. Removing it from there.

Extending the documentation for SdrNodeContext and SdrNodeRole via the GetRole docs.

(Internal change: 2002225)