Add constexpr copy constructor to half.
This allows half (aka GfHalf) values to be used as
arguments to constexpr constructors. This is not useful
by itself since half can't be constexpr constructed,
but it allows us to create consistent constexpr
interfaces for the GfVec* types. In the future, we
may look at making half constexpr-constructible (at
least for some special values like 0.0 and 1.0).

(Internal change: 2003181)