Vstruct support in HdPrman.
This adds a mini material filtering ("matfilt") system,
which includes a resolver for vstructs.  In the future this may
move to another level of the system; for now it gives us
improved correctness and capability for Renderman specific
materials.

This does add a conversion step in and out of the matfilt
rep; pending Hydra material network work we expect this
to go away or reduce in the future.

The new debug code HDPRMAN_VSTRUCT provides diagnostics
for vstructs specifically.

(Internal change: 2005334)
(Internal change: 2005462)
(Internal change: 2005540)
(Internal change: 2005638)
(Internal change: 2006059)
(Internal change: 2006105)