Pass original MObject with MFnMeshData around wherever mesh.object() is needed for iteration.
I think the problem is that data of type kMeshData that is flowing through
the dependency graph does not exactly match Maya's RTTI, and as a result
mesh.object() returns nil, even though the original MObject used to create
the MFnMesh function set is still valid - there doesn't seem to be any other
way to get that MObject back either. This PR just replaces the arguments of
type MFnMesh for methods that need the MObject later (because they use
MItMeshFaceVertex) with the original MObject the MFnMesh was constructed
from.

I also believe this bug has larger implications than just normals - the code
should also break for UVs and mesh colors when the mesh has skeletal
animation (but I have not verified this).

(Internal change: 2006888)