Fix perf regression in batch rebuild.
Our materialTag isn't empty by default but instead uses 'defaultMaterialTag'.
The mistake in the code was to start each Sync with an 'empty' material tag and compare it to the materialTag we found for the material.
But when the shader is not dirty, we do not update the materialTag.
Meaning we would compare 'empty' with 'defaultMaterialTag' every frame, causing batches to rebuild every frame.

Same applied to comparing the shader sources. (only during dirty surface do we update those strings)

(Internal change: 2007230)