Implement vis/invis commands directly in AppController rather than backwardly calling on UI elements to effect the mutation, especially since those UI elements may no longer be initialized at calling time.
Remove all calls to _resetPrimView and _resetPrimViewVis from
visibility-mutation code, as it is entirely redundant since making
the GUI responsive to stage changes.  In future, when the
rootDataModel sends out more granular signals about changes,
we can make the PrimViewWidget itself institute the cheaper
updates when applicable.

Add a blackBoxTesting.md to usdviewq, to record test plans
for features it's difficult to test in the automated testusdview
harness.

(Internal change: 2007395)