[usdImaging] Fix over-invalidation on instancer property edits.
Due to abuse of the primInfo datastructure, we were registering USD prim dependencies between USD instancers and hydra prototypes.  Due to limitations of the USD Imaging edit API, we weren't able to distinguish between prototype and instancer updates.

This was (for example) causing us to resync a point instancer every time it got an edit for "positions".

(Internal change: 2011208)