[usdImaging] Filter out resync/populate calls inside USD masters.
In one reported case, for example, resyncing a model prim would resync all of the (singly-instantiated) masters representing instances in the model.  Since each resync call ends up as a call to populate (in case new prims were added), we'd get a resync on both the instance and the master, and double-insert prims, some of which wouldn't have the benefit of their instance transform or bound material.

This matches behavior before the USD edit change, although the check is in a more appropriate place now.

Fixes #1003
Fixes #1005

(Internal change: 2013847)