[build] fix check for PATH env variable settings given to --env-var option
The intention is for the PATH environment variable *not* to be modified
directly and instead be adjusted using the --pre-path or --post-path
options. The check for this though was incorrectly testing the full
"key=value" string as opposed to just the key.

This issue was identified and addressed by Autodesk here:
https://github.com/Autodesk/maya-usd/pull/109

(Internal change: 2021260)