Fix testSdrRegistry failure with monolithic builds
The addition of the .glslfx parser plugin in
change 2019889 caused this failure, as the test
was not expecting to find a parser that could
handle .glslfx files. This plugin wouldn't be
found in normal builds because the test disables
the standard plugin search path, but in monolithic
builds the plugin is part of the library and gets
picked up by the test.

A new env var was added to disable parser plugin
discovery, and a new method was added to allow
the test to explicitly register the parser plugins
it wants to use. This matches how we deal with
discovery plugins in this test.

(Internal change: 2021443)