Fix issues seen when rendering test models in Houdini/Solaris. We now make two texture calls, once for color and once to look up an alpha.
Before this change, if a texture did not have an alpha channel, alpha
would be 0, causing the test models in Solaris to look like glass
when rendering with hdPrman. Note that the "missingalpha" parameter
in for the texture call in OSL is only used if the texture file itself
is missing, not when the texture does not contain an alpha channel.

(Internal change: 2023822)