Prevent double color correction on MacOS by disabling color quantization in ColorizeTask by default.
This works because when a platform is 'color correction capable' we always enable color quantization by default (because we call SetColorCorrectionParams).
So all platforms that previously relied on this being enabled, will continue to have it enabled.

MacOS is currently NOT 'colorcorrectioncapable' so no call to SetColorCorrectionParams will enable quantization.
MacOS will then rely on setting SRGB in openGL for the linear to srgb transform.

(Internal change: 2029075)