Storm Volumes: stop ray marching when hitting opaque geometry.
The shader volume.glslfx reads the depth buffer in computeRayEndDepth to stop ray marching in volumeIntegrator. Note that this only works correctly when Storm is configured to use AOVs.

In detail:

The HdxOitVolumeRenderTask requests the HdStRenderPassShader to bind the texture baking the depth AOV.

The result of HdStRenderPassShader::GetParams() prompts binding resolution reserve a sampler unit for this texture and makes codegen to produce an accessor HdGet_depthReadback() that the shader can use. HdStRenderPassShader::BindResources then gets the texture from the depth AOV to bind it to the sampler unit. To tie this together, it also writes the sampler unit to the uniform sampler2D accessed by HdGet_depthReadback().

(Internal change: 2030308)