Changing the behavior of ComputePurpose() in UsdGeomImageable to no longer be pruning and instead behave as follows:
A prim that is Imageable with an authored opinion about its purpose will
always have the same effective purpose as its authored value.  If the
prim is not Imageable or does not have an authored opinion about its own
purpose, then it will inherit the purpose of the closest Imageable
ancestor with an authored purpose opinion.  If there are no Imageable
ancestors with an authored purpose opinion then this prim uses its
fallback purpose.

This also updates usdImaging and bounding box caching to reflect the new
computation of effective purpose.

(Internal change: 2033434)